fix(mobile): corregir crashes de arranque y documentar pasos para correr la app
La app crasheaba al abrir (verificado instalandola en un emulador y
revisando logcat), por dos bugs reales en la navegacion type-safe:

1. AlertaPerdidaDetalle (routes.kt) tenia campos latitud/longitud de tipo
   Double. Navigation Compose no puede inferir un NavType para Double
   automaticamente (solo Int/Long/Boolean/Float/String y variantes), asi
   que registrar ese destino lanzaba IllegalArgumentException apenas se
   construia el NavHost - la app crasheaba en el primer frame, siempre.
   Cambiados a Float (precision de sobra para un pin en el mapa),
   convirtiendo en los dos bordes donde se usan (AppNavigation.kt).

2. OnboardingScreen.kt navegaba con navController.navigate("Login")
   (ruta de texto), pero el grafo de navegacion usa rutas tipadas
   (`object Login` serializable). Object Login no coincide con el string
   "Login" en el grafo generado, asi que terminar el onboarding
   crasheaba con "Navigation destination ... cannot be found". Cambiado a
   navigate(Login) (tipado, con el import correspondiente).

Verificado instalando el APK en un emulador real: onboarding -> permisos
de notificacion -> login, sin crashes, con capturas de pantalla.

De paso se incluye el trabajo ya presente sin commitear en MainActivity.kt
(limpieza de imports, consistencia de SharedPreferences entre
SessionManager y UserRoleViewModel, desconectar el WebSocket cuando no
hay usuario) - no causaba el crash pero estaba probado y funcionando
junto con este fix.

README actualizado con pasos exactos para correr la app y una nota de
troubleshooting sobre estos dos patrones de crash.

+++ app/src/main/java/com/proyectopoo/petcareapp/navigation/AppNavigation.kt
@@ -1255,8 +1255,8 @@ fun AppNavigation(
                                 petsId = alerta.petsId,
                                 reporterId = alerta.usuarioId,
                                 descripcion = alerta.descripcion,
-                                latitud = alerta.latitud,
-                                longitud = alerta.longitud,
+                                latitud = alerta.latitud.toFloat(),
+                                longitud = alerta.longitud.toFloat(),
                                 direccionTexto = alerta.direccionTexto,
                                 estado = alerta.estado ?: "ACTIVA",
                                 currentUserId = args.usuarioId
@@ -1272,8 +1272,8 @@ fun AppNavigation(
             AlertaPerdidaDetalleScreen(
                 alertaId = args.alertaId,
                 descripcion = args.descripcion,
-                latitud = args.latitud,
-                longitud = args.longitud,
+                latitud = args.latitud.toDouble(),
+                longitud = args.longitud.toDouble(),
                 direccionTexto = args.direccionTexto,
                 estadoInicial = args.estado,
                 isReporter = args.reporterId == args.currentUserId,
